mobile: split reaction popover helpers
